june 22 updates

+++ Amazon/src/main/java/utils/Base.java
@@ -2,8 +2,10 @@ package utils;
 
 import java.util.concurrent.TimeUnit;
 
+//import org.openqa.selenium.Dimension;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.firefox.FirefoxDriver;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
 import com.relevantcodes.extentreports.ExtentReports;
@@ -11,26 +13,42 @@ import com.relevantcodes.extentreports.ExtentTest;
 
 public class Base {
 	
-	public static String baseUrl = "https://amazon.com";
+	public static String strBaseUrl;
+	public static String strBrowser;
 	public static WebDriver driver;
 	public static WebDriverWait wait;
-	public static ExtentReports report = new ExtentReports("D:/Automation/Reports/ExtentReport.html");
-	public static ExtentTest test = report.startTest("ReportDemo");
+	public static ExtentReports report;
+	public static ExtentTest test;
 	
 	public void launchBrowser() throws Exception
 	{
 		try {
-			 Runtime.getRuntime().exec("taskkill /F /IM chromedriver.exe /T");
-			 Runtime.getRuntime().exec("taskkill /F /IM chrome.exe /T");
-			 Runtime.getRuntime().exec("taskkill /F /IM geckodriver.exe /T");
-			 Runtime.getRuntime().exec("taskkill /F /IM firefox.exe /T");
+			 report = new ExtentReports(System.getProperty("user.dir")+"\\ExtentReport\\ExtentReportResults.html");
+			 test = report.startTest("ExtentDemo");
+			 strBrowser = Config.readPropertyFile("Browser");
 			 
-			 System.setProperty("webdriver.firefox.bin", "C:\\Users\\vinoth.b\\AppData\\Local\\Mozilla Firefox\\firefox.exe");
+			 if(strBrowser.equalsIgnoreCase("Chrome"))
+			 {
+				 Runtime.getRuntime().exec("taskkill /F /IM chromedriver.exe /T");
+				 Runtime.getRuntime().exec("taskkill /F /IM chrome.exe /T");
+				 driver = new ChromeDriver();
+				 driver.manage().window().maximize();
+				 driver.manage().timeouts().implicitlyWait(3, TimeUnit.SECONDS);
+				 wait = new WebDriverWait(driver, 10);
+			 }
 			 
-			 driver = new ChromeDriver();
-			 driver.manage().window().maximize();
-			 driver.manage().timeouts().implicitlyWait(3, TimeUnit.SECONDS);
-			 wait = new WebDriverWait(driver, 10);
+			 if(strBrowser.equalsIgnoreCase("Firefox"))
+			 {
+				 Runtime.getRuntime().exec("taskkill /F /IM geckodriver.exe /T");
+				 Runtime.getRuntime().exec("taskkill /F /IM firefox.exe /T");
+				 System.setProperty("webdriver.firefox.bin", "C:\\Users\\vinoth.b\\AppData\\Local\\Mozilla Firefox\\firefox.exe");
+				 driver = new FirefoxDriver();
+				 driver.manage().window().maximize();
+				 driver.manage().timeouts().implicitlyWait(3, TimeUnit.SECONDS);
+				 wait = new WebDriverWait(driver, 10);
+			 }
+			 
+			 
 		}
 		catch(Exception e)
 		{
@@ -42,7 +60,8 @@ public class Base {
 	
 	public void launchApp() throws Exception
 	{
-		 driver.get(baseUrl);
+		 strBaseUrl = Config.readPropertyFile("URL");
+		 driver.get(strBaseUrl);
 		}
 		 
 	
@@ -50,6 +69,9 @@ public class Base {
 	{
 		 driver.close();
 		 driver.quit();
+		 report.endTest(test);
+		 report.flush();
+//		 report.close();
 		}
